[6868] Extract the domain part of the image bounded context in a dedicated module
Bug: https://github.com/eclipse-sirius/sirius-web/issues/6868
