+++ packages/images/backend/sirius-web-images-domain/.settings/org.springframework.ide.eclipse.prefs
@@ -1,0 +1,3 @@
+boot.validation.initialized=true
+eclipse.preferences.version=1
+
