+++ packages/images/backend/sirius-web-images-domain/src/main/java/org/eclipse/sirius/web/images/domain/services/ImageDeletionService.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-package org.eclipse.sirius.web.domain.boundedcontexts.image.services;
+package org.eclipse.sirius.web.images.domain.services;
 
 import java.util.Objects;
 import java.util.UUID;
@@ -19,9 +19,9 @@ import org.eclipse.sirius.components.events.ICause;
 import org.eclipse.sirius.web.core.domain.results.Failure;
 import org.eclipse.sirius.web.core.domain.results.IResult;
 import org.eclipse.sirius.web.core.domain.results.Success;
-import org.eclipse.sirius.web.domain.boundedcontexts.image.repositories.IImageRepository;
-import org.eclipse.sirius.web.domain.boundedcontexts.image.services.api.IImageDeletionService;
 import org.eclipse.sirius.web.domain.services.api.IMessageService;
+import org.eclipse.sirius.web.images.domain.repositories.IImageRepository;
+import org.eclipse.sirius.web.images.domain.services.api.IImageDeletionService;
 import org.springframework.stereotype.Service;
 
 /**
