+++ packages/releng/backend/sirius-components-test-coverage/pom.xml
@@ -448,6 +448,11 @@
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.sirius</groupId>
+			<artifactId>sirius-web-images-domain</artifactId>
+			<version>2026.7.3</version>
+		</dependency>
+		<dependency>
+			<groupId>org.eclipse.sirius</groupId>
 			<artifactId>sirius-web-application</artifactId>
 			<version>2026.7.3</version>
 		</dependency>
