+++ packages/sirius-web/backend/sirius-web-application/pom.xml
@@ -49,6 +49,11 @@
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.sirius</groupId>
+			<artifactId>sirius-web-images-domain</artifactId>
+			<version>2026.7.3</version>
+		</dependency>
+		<dependency>
+			<groupId>org.eclipse.sirius</groupId>
 			<artifactId>sirius-components-graphql-api</artifactId>
 			<version>2026.7.3</version>
 		</dependency>
