+++ packages/sirius-web/backend/sirius-web/src/test/java/org/eclipse/sirius/web/application/services/CustomImageMetadataSearchServiceTests.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2024, 2025 Obeo.
+ * Copyright (c) 2024, 2026 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -21,9 +21,9 @@ import org.eclipse.sirius.components.view.emf.api.CustomImageMetadata;
 import org.eclipse.sirius.components.view.emf.api.ICustomImageMetadataSearchService;
 import org.eclipse.sirius.web.AbstractIntegrationTests;
 import org.eclipse.sirius.web.data.StudioIdentifiers;
-import org.eclipse.sirius.web.domain.boundedcontexts.image.services.api.IImageCreationService;
-import org.eclipse.sirius.web.domain.boundedcontexts.image.services.api.IImageDeletionService;
-import org.eclipse.sirius.web.domain.boundedcontexts.image.services.api.IImageSearchService;
+import org.eclipse.sirius.web.images.domain.services.api.IImageCreationService;
+import org.eclipse.sirius.web.images.domain.services.api.IImageDeletionService;
+import org.eclipse.sirius.web.images.domain.services.api.IImageSearchService;
 import org.eclipse.sirius.web.tests.data.GivenSiriusWebServer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
